Fixed A Gametexture bug that adds Frames to a tex

+++ include/entity.h
@@ -39,7 +39,7 @@ class Entity {
   void ResetFrame();
   void SetFrameToStartPos();
 
-  short Frame;
+  int Frame;
   Event<> OnTextureSwitch;
   Vector2 Position;
   GameTexture CurrentTexture;
